object_detection_app/requirements.txt
streamlit==1.32.0
tensorflow-cpu==2.15.0
opencv-python-headless==4.9.0.80
pillow==10.2.0
numpy==1.26.4
urllib3==2.0.7 